changes to ranged weapon class
All weapons now have 2 attacks, one with right and one with left click. A bunch of stuff was moved from the individual weapon to the ranged weapon class, meaning you have to write less code to make a new weapon. A new weapon was created. For examples, look in the ranged weapon scripts. To test new weapons, remove the current weapon from the weapon holder on the player, and simply drag and drop a weapon prefab onto the weapon holder to replace it.